RBC Model With Different Rho Values
An experiment looking to model the differences in impulse responses to a shock of 1% to total factor productivity when different values for rho are present, which represents the lasting power of the prior term's TFP in the TFP evolution formula. TFP in this model follows an autoregressive process, and thus we wish to examine how impulse responses to labor, consumption, and output change over time for different values for this parameter.